+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - scripts to deploy stage 0: Auction
 - additional helpers for `Auction` and `Rewards` in MultiSig
 - scripts to deploy stage 1: DeFi farming
+- stop method for farming
 
 ### Changed
 - merge `BurnableToken` and `MaxAllowedSupply` into `MintableToken`